requirements.txt
playwright==1.54.0
python-dotenv==1.0.1
supabase==2.7.4
pandas==2.2.2
openpyxl==3.1.5
pyarrow==16.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
